requirements.txt
pip==25.1.1
psycopg2-binary==2.9.10
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
celery==5.3.6
defusedxml==0.7.1
Django==4.2.20
djangorestframework==3.16.0
django_environ==0.12.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
kombu==5.5.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-toolbelt==1.0.0
redis==5.2.1
rich==14.0.0
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
drf_spectacular==0.28.0
uritemplate==4.1.1
PyYAML==6.0.2
jsonschema==4.23.0
inflection==0.5.1
attrs==25.3.0
jsonschema_specifications==2025.4.1
rpds_py==0.24.0
typing_extensions==4.13.2
referencing==0.36.2

